increased height of results list to show first five items
moved specific results list styles to SearchResultsList because they were only used once in SelectList
