+++ src/components/Search/SearchResultsList.tsx
@@ -1,8 +1,15 @@
 import Box from '@mui/material/Box';
 import List from '@mui/material/List';
+import { style } from 'typestyle';
 
 import { useAppSelector } from '../../hooks/app';
-import { SelectListStyles } from '../../styles';
+import {
+  Colors,
+  FontsRaw,
+  Spacing,
+  captionStylesLight,
+  captionStylesDark,
+} from '../../styles';
 import { Changeset } from '../../types/state';
 import SearchResultsListItem from './SearchResultsListItem';
 
@@ -14,6 +21,89 @@ interface SearchResultsListProps {
   listItemComponent?: 'checkbox' | 'radio';
 }
 
+const sharedSelectStyles = {
+  borderRadius: '4px',
+  marginTop: `${Spacing.xSmall}px`,
+  height: '285px',
+  overflow: 'auto',
+  maxWidth: '100%',
+  padding: `${Spacing.xSmall}px`,
+  border: `1px solid ${Colors.BorderDefault}`,
+};
+
+const SelectListLight = style({
+  backgroundColor: Colors.Background300,
+  zIndex: 100,
+  position: 'relative',
+  ...sharedSelectStyles,
+  $nest: {
+    '.MuiListItemButton-root': {
+      padding: `${Spacing.xSmall}px ${Spacing.Small}px`,
+      $nest: {
+        '&:hover': {
+          backgroundColor: Colors.SecondaryHover,
+          borderRadius: '4px',
+        },
+        '&:active': {
+          backgroundColor: Colors.SecondaryActive,
+          borderRadius: '4px',
+        },
+      },
+    },
+    '.item-selected': {
+      backgroundColor: Colors.SecondaryHover,
+      borderRadius: '4px',
+    },
+    '.revision-hash': {
+      ...FontsRaw.BodyDefault,
+      marginRight: Spacing.Small,
+    },
+    '.info-caption': {
+      ...captionStylesLight,
+    },
+    '.MuiTypography-root': {
+      ...FontsRaw.BodyDefault,
+    },
+  },
+});
+
+const SelectListDark = style({
+  backgroundColor: Colors.Background300Dark,
+  zIndex: 100,
+  position: 'relative',
+  ...sharedSelectStyles,
+  $nest: {
+    '.MuiListItemButton-root': {
+      padding: `${Spacing.xSmall}px ${Spacing.Small}px`,
+
+      $nest: {
+        '&:hover': {
+          backgroundColor: Colors.SecondaryHoverDark,
+          borderRadius: '4px',
+        },
+        '&:active': {
+          backgroundColor: Colors.SecondaryActiveDark,
+          borderRadius: '4px',
+        },
+      },
+    },
+    '.item-selected': {
+      backgroundColor: Colors.SecondaryHoverDark,
+      borderRadius: '4px',
+    },
+    '.revision-hash': {
+      ...FontsRaw.BodyDefault,
+      marginRight: Spacing.Small,
+    },
+    '.info-caption': {
+      ...captionStylesDark,
+    },
+    '.MuiTypography-root': {
+      ...FontsRaw.BodyDefaultDark,
+    },
+  },
+});
+
 function SearchResultsList({
   compact,
   searchResults,
@@ -22,7 +112,7 @@ function SearchResultsList({
   listItemComponent,
 }: SearchResultsListProps) {
   const mode = useAppSelector((state) => state.theme.mode);
-  const styles = SelectListStyles(mode);
+  const styles = mode === 'light' ? SelectListLight : SelectListDark;
 
   const isInProgressChecked = (item: Changeset) => {
     return displayedRevisions.map((rev) => rev.id).includes(item.id);
